Tracking/requirements.txt
matplotlib
moviepy
numpy
Pillow
scikit-image
pytorch
torchvision
tqdm
pytest
